Reuse copy-bookmarklet script
This change moves the inline script used for copying text to the clipboard into a separate file and adds it to the page as necessary. It also refactors that script very slightly, and fixes one use of the script that was not working properly—seems like maybe the code got reformatted as if it was a block of plain text, which broke some of the code.